# improvement(planner): hash join ~6x performance improvement