--- a/query/Cargo.toml
+++ b/query/Cargo.toml
@@ -116,6 +116,7 @@ twox-hash = "1.6.2"
 typetag = "0.1.8"
 uuid = { version = "0.8.2", features = ["serde", "v4"] }
 walkdir = "2.3.2"
+smallvec = { version = "1.8.0", features = ["write"] }
 
 [dev-dependencies]
 clickhouse-driver = { git = "https://github.com/datafuse-extras/clickhouse_driver", rev = "cf978da" }