--- a/query/src/common/hashtable/hash_table.rs
+++ b/query/src/common/hashtable/hash_table.rs
@@ -35,13 +35,19 @@ pub struct HashTable<Key: HashTableKeyable, Entity: HashTableEntity<Key>, Grower
     generics_hold: PhantomData<Key>,
 }
 
-unsafe impl<Key: HashTableKeyable, Entity: HashTableEntity<Key>, Grower: HashTableGrower> Send
-    for HashTable<Key, Entity, Grower>
+unsafe impl<
+        Key: HashTableKeyable + Send,
+        Entity: HashTableEntity<Key> + Send,
+        Grower: HashTableGrower,
+    > Send for HashTable<Key, Entity, Grower>
 {
 }
 
-unsafe impl<Key: HashTableKeyable, Entity: HashTableEntity<Key>, Grower: HashTableGrower> Sync
-    for HashTable<Key, Entity, Grower>
+unsafe impl<
+        Key: HashTableKeyable + Sync,
+        Entity: HashTableEntity<Key> + Sync,
+        Grower: HashTableGrower,
+    > Sync for HashTable<Key, Entity, Grower>
 {
 }