--- a/query/src/pipelines/new/processors/transforms/hash_join/mod.rs
+++ b/query/src/pipelines/new/processors/transforms/hash_join/mod.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 mod chaining_hash_table;
-mod hash;
 mod hash_table;
 mod row;