--- a/query/src/pipelines/new/processors/transforms/hash_join/row.rs
+++ b/query/src/pipelines/new/processors/transforms/hash_join/row.rs
@@ -12,20 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::sync::Arc;
 use std::sync::RwLock;
 
 use common_datablocks::DataBlock;
-use common_datavalues::ColumnRef;
-use common_datavalues::DataField;
-use common_datavalues::DataSchema;
 use common_datavalues::DataSchemaRef;
 use common_exception::Result;
-use common_functions::scalars::FunctionContext;
-use common_planners::Expression;
 use smallvec::SmallVec;
-
-use crate::common::ExpressionEvaluator;
 
 pub type KeysVector = Vec<SmallVec<[u8; 16]>>;
 
@@ -106,63 +98,4 @@ impl RowSpace {
     fn gather_single_chunk(&self, chunk: &Chunk, indices: &[u32]) -> Result<DataBlock> {
         DataBlock::block_take_by_indices(&chunk.data_block, indices)
     }
-}
-
-// TODO(leiysky): compare in a more efficient way
-pub fn compare_and_combine(
-    func_ctx: &FunctionContext,
-    probe_input: DataBlock,
-    probe_result: DataBlock,
-    build_keys: &[ColumnRef],
-    probe_keys: &[ColumnRef],
-) -> Result<DataBlock> {
-    assert_eq!(build_keys.len(), probe_keys.len());
-    let mut compare_exprs: Vec<Expression> = Vec::with_capacity(build_keys.len());
-    let mut data_fields: Vec<DataField> = Vec::with_capacity(build_keys.len() + probe_keys.len());
-    let mut columns: Vec<ColumnRef> = Vec::with_capacity(build_keys.len() + probe_keys.len());
-    for (idx, (build_key, probe_key)) in build_keys.iter().zip(probe_keys).enumerate() {
-        let build_key_name = format!("build_key_{idx}");
-        let probe_key_name = format!("probe_key_{idx}");
-
-        let build_key_data_type = build_key.data_type();
-        let probe_key_data_type = probe_key.data_type();
-
-        columns.push(build_key.clone());
-        columns.push(probe_key.clone());
-
-        data_fields.push(DataField::new(build_key_name.as_str(), build_key_data_type));
-        data_fields.push(DataField::new(probe_key_name.as_str(), probe_key_data_type));
-
-        let compare_expr = Expression::BinaryExpression {
-            left: Box::new(Expression::Column(build_key_name.clone())),
-            right: Box::new(Expression::Column(probe_key_name.clone())),
-            op: "=".to_string(),
-        };
-        compare_exprs.push(compare_expr);
-    }
-
-    let predicate = compare_exprs
-        .into_iter()
-        .reduce(|prev, next| Expression::BinaryExpression {
-            left: Box::new(prev),
-            op: "and".to_string(),
-            right: Box::new(next),
-        })
-        .unwrap();
-
-    let data_block = DataBlock::create(Arc::new(DataSchema::new(data_fields)), columns);
-
-    let filter = ExpressionEvaluator::eval(func_ctx, &predicate, &data_block)?;
-
-    let mut produce_block = probe_input;
-    for (col, field) in probe_result
-        .columns()
-        .iter()
-        .zip(probe_result.schema().fields().iter())
-    {
-        produce_block = produce_block.add_column(col.clone(), field.clone())?;
-    }
-    produce_block = DataBlock::filter_block(&produce_block, &filter)?;
-
-    Ok(produce_block)
 }