--- a/query/src/pipelines/transforms/group_by/mod.rs
+++ b/query/src/pipelines/transforms/group_by/mod.rs
@@ -21,7 +21,7 @@ mod aggregator_polymorphic_keys;
 mod aggregator_state;
 mod aggregator_state_entity;
 mod aggregator_state_iterator;
-mod keys_ref;
+pub(crate) mod keys_ref;
 
 pub use aggregator::Aggregator;
 pub use aggregator_groups_builder::GroupColumnsBuilder;