@xudong963 2022-05-21
> Is the flame graph generated from a release build?

No, from debug build.  The performance results are from release build