@sundy-li 2022-05-23
This SQL can reproduce the bug with a smaller data set, you can use this to debug.
```
select count(*) from (select to_int32(a) as a from t1 order by a limit 10) t1 inner join (select to_int32(a) as a from t2 order by a limit 10) t2 on t1.a=t2.a
```